requirements.txt
Django==3.2.5
django_filter==21.1
python-dotenv==0.20.0
